Fix toplevel
Was going past the toplevel. We can't patch the difference between
the max const degree and the const degree for this output wire on the
zhats since there are n of them! My fix: simply multiply the output wire
by the necessary number of vhat encodings to raise up the needed Y
index.